[core] Extract RayEventRecorderBase from RayEventRecorder
